refactor: 우숭자에 대한 테스트를 CarsTest에서 WinnersTest로 이동
