+++ src/test/java/racingcar/domain/CarsTest.java
@@ -3,9 +3,6 @@ package racingcar.domain;
 import static org.assertj.core.api.Assertions.*;
 import static racingcar.domain.Cars.*;
 
-import java.util.Arrays;
-import java.util.List;
-
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -35,29 +32,5 @@ class CarsTest {
 		Assertions.assertThatThrownBy(() -> new Cars(Utils.splitter("pobi,pobi")))
 			.isInstanceOf(IllegalArgumentException.class)
 			.hasMessageContaining(ERROR_DUPLICATE_NAME);
-	}
-
-	@Test
-	@DisplayName("우승자가 한 명인 경우 테스트")
-	void getOneWinner() {
-		List<Car> carList = Arrays.asList(
-			new Car("pobi", 2),
-			new Car("crong", 3),
-			new Car("honux", 1)
-		);
-		Cars cars = new Cars(carList);
-		assertThat(cars.getWinners()).contains("crong");
-	}
-
-	@Test
-	@DisplayName("우승자가 여러명인 경우 테스트")
-	void getSeveralWinner() {
-		List<Car> carList = Arrays.asList(
-			new Car("pobi", 3),
-			new Car("crong", 3),
-			new Car("honux", 1)
-		);
-		Cars cars = new Cars(carList);
-		assertThat(cars.getWinners().stream()).contains("pobi", "crong");
 	}
 }
